requirements.txt
asgiref==3.3.1
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-countries==7.1
django-environ==0.4.5
django-filter==2.4.0
django-phonenumber-field==5.0.0
django-redis-cache==3.0.0
dj-rest-auth==2.1.4 
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
importlib-metadata==3.7.2
Markdown==3.3.4
oauthlib==3.1.0
phonenumbers==8.12.20
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-json-logger==2.0.1
python3-logstash==0.4.80
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
zipp==3.4.1
